backend/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
argilla==2.2.2
asttokens==2.4.1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
cryptography==43.0.1
cssselect==1.2.0
dataclasses-json==0.6.7
datasets==3.0.1
debugpy==1.8.6
decorator==5.1.1
deepdiff==8.0.1
dill==0.3.8
dnspython==2.6.1
docopt==0.6.2
emoji==2.13.2
et-xmlfile==1.1.0
eutils==0.6.0
executing==2.1.0
filelock==3.16.1
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.6.1
google==3.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
habanero==1.2.6
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.6
langchain-google-genai==2.0.0
langchain-mongodb==0.2.0
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.129
Levenshtein==0.26.0
lxml==5.3.0
lxml_html_clean==0.2.2
Markdown==3.7
markdown-it-py==3.0.0
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
metapub==0.5.12
multidict==6.1.0
multiprocess==0.70.16
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openpyxl==3.1.5
orderly-set==5.2.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.9.1
pypandoc==1.13
pyparsing==3.1.4
pypdf==5.0.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-Levenshtein==0.26.0
python-magic==0.4.27
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.5.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
unstructured==0.15.13
unstructured-client==0.25.9
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
websockets==13.1
wrapt==1.16.0
XlsxWriter==3.2.0
xmltodict==0.13.0
xxhash==3.5.0
yarl==1.13.1
yt-dlp==2024.9.27
